+++ MapDriveInator/scripts/CreateBaseFolderNode.groovy
@@ -4,7 +4,12 @@ import MDI
 baseFolderNode = MDI.obtainBaseFolder(node)
 if(baseFolderNode){
     c.select(baseFolderNode)
-	baseFolderNode.style.name = 'baseFolder'
+    try {
+	 baseFolderNode.style.name = MDI.styleBaseFolder
+    } catch(Exception ex) {
+        ui.showMessage("The mindmap has no '${MDI.styleBaseFolder}' style.\n\nPlease import the MDI styles into your map.", 0)
+        return 0
+    }
     def nameFilt = MDI.getFilter(baseFolderNode)
     def maxD = MDI.getMaxDepth(baseFolderNode)
 }else{