+++ MapDriveInator/history.md
@@ -1,5 +1,12 @@
 # History
 
+## v0.0.5
+
+* corrected createFilesFromNotes
+* Added 'Import MDI styles'
+(without library: only works for users that installed other add-ons that have the `edofro.pseudofreeplaneapi.UserStyles` library.)
+* Added exception handling when the mindmap doesn't have MDI styles
+
 ## v0.0.4
 
 * Added Template map